+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.2
+- Moved `bsp_var_list` from on-core memory to external memory
+- Host no longer interferes with variable registrattion
 - Added `bsp_get`, `bsp_hpget` and `bsp_put`
 - Fixed timer to use real time instead of clockcycles
 - Added `ebsp_message`